snp_genotypes, is_accessible
