+++ malariagen_data/anopheles.py
@@ -38,7 +38,6 @@ from .util import (
     DIM_VARIANT,
     CacheMiss,
     Region,
-    da_compress,
     da_from_zarr,
     dask_compress_dataset,
     hash_params,
@@ -1388,135 +1387,6 @@ class AnophelesDataResource(
             sample_query = np.nonzero(loc_samples)[0].tolist()
 
         return sample_sets, sample_query
-
-    def _snp_genotypes_for_contig(
-        self,
-        *,
-        contig: str,
-        sample_set: str,
-        field: str,
-        inline_array: bool,
-        chunks: str,
-    ) -> da.Array:
-        """Access SNP genotypes for a single contig and a single sample set."""
-        assert isinstance(contig, str)
-        assert isinstance(sample_set, str)
-        root = self.open_snp_genotypes(sample_set=sample_set)
-        z = root[f"{contig}/calldata/{field}"]
-        d = da_from_zarr(z, inline_array=inline_array, chunks=chunks)
-
-        return d
-
-    @doc(
-        summary="Access SNP genotypes and associated data.",
-        returns="""
-            An array of either genotypes (GT), genotype quality (GQ), allele
-            depths (AD) or mapping quality (MQ) values.
-        """,
-    )
-    def snp_genotypes(
-        self,
-        region: base_params.region,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        sample_query: Optional[base_params.sample_query] = None,
-        field: base_params.field = "GT",
-        site_mask: Optional[base_params.site_mask] = None,
-        inline_array: base_params.inline_array = base_params.inline_array_default,
-        chunks: base_params.chunks = base_params.chunks_default,
-    ) -> da.Array:
-        debug = self._log.debug
-
-        debug("normalise parameters")
-        sample_sets = self._prep_sample_sets_param(sample_sets=sample_sets)
-        resolved_region = self.resolve_region(region)
-        del region
-
-        debug("normalise region to list to simplify concatenation logic")
-        if isinstance(resolved_region, Region):
-            resolved_region = [resolved_region]
-
-        debug("concatenate multiple sample sets and/or contigs")
-        lx = []
-        for r in resolved_region:
-            ly = []
-
-            for s in sample_sets:
-                y = self._snp_genotypes_for_contig(
-                    contig=r.contig,
-                    sample_set=s,
-                    field=field,
-                    inline_array=inline_array,
-                    chunks=chunks,
-                )
-                ly.append(y)
-
-            debug("concatenate data from multiple sample sets")
-            x = da.concatenate(ly, axis=1)
-
-            debug("locate region - do this only once, optimisation")
-            if r.start or r.end:
-                pos = self.snp_sites(region=r.contig, field="POS")
-                loc_region = locate_region(r, pos)
-                x = x[loc_region]
-
-            lx.append(x)
-
-        debug("concatenate data from multiple regions")
-        d = da.concatenate(lx, axis=0)
-
-        debug("apply site filters if requested")
-        if site_mask is not None:
-            loc_sites = self.site_filters(
-                region=resolved_region,
-                mask=site_mask,
-            )
-            d = da_compress(loc_sites, d, axis=0)
-
-        debug("apply sample query if requested")
-        if sample_query is not None:
-            df_samples = self.sample_metadata(sample_sets=sample_sets)
-            loc_samples = df_samples.eval(sample_query).values
-            d = da.compress(loc_samples, d, axis=1)
-
-        return d
-
-    @doc(
-        summary="Compute genome accessibility array.",
-        returns="An array of boolean values identifying accessible genome sites.",
-    )
-    def is_accessible(
-        self,
-        region: base_params.region,
-        site_mask: base_params.site_mask = DEFAULT,
-    ) -> np.ndarray:
-        debug = self._log.debug
-
-        debug("resolve region")
-        resolved_region = self.resolve_region(region)
-        del region
-
-        debug("determine contig sequence length")
-        seq_length = self.genome_sequence(resolved_region).shape[0]
-
-        debug("set up output")
-        is_accessible = np.zeros(seq_length, dtype=bool)
-
-        pos = self.snp_sites(region=resolved_region, field="POS").compute()
-        if resolved_region.start:
-            offset = resolved_region.start
-        else:
-            offset = 1
-
-        debug("access site filters")
-        filter_pass = self.site_filters(
-            region=resolved_region,
-            mask=site_mask,
-        ).compute()
-
-        debug("assign values from site filters")
-        is_accessible[pos - offset] = filter_pass
-
-        return is_accessible
 
     def _snp_df(self, *, transcript: str) -> Tuple[Region, pd.DataFrame]:
         """Set up a dataframe with SNP site and filter columns."""
